// src/app/core/interfaces/perfil.ts
export interface Perfil {
    idPerfil?: number;
    idCia?: number;
    idCO?: string;
    nombreCompleto?: string;
    documentoIdentidad?: string;
    telefono?: string;
    direccion?: string;
    email?: string;
    idMunicipio?:number;
    municipio?: string;
    usuario?: string;
    imagen?: string;
}